factor loop / recursion handling out into AELoopRecursion.cpp
AbstractInterpretation.cpp had grown to ~1700 lines.  Move the
self-contained loop / recursion module into a new
AELoopRecursion.cpp:

  * handleLoopOrRecursion (widen/narrow fixpoint driver)
  * Dense base cycle helpers — getFullCycleHeadState,
    widenCycleState, narrowCycleState
  * Recursion utilities — isRecursiveFun, isRecursiveCallSite,
    skipRecursiveCall, skipRecursionWithTop, shouldApplyNarrowing

These all share the same conceptual scope (cycles + recursion-as-cycle
+ widening/narrowing termination) and were sprinkled across AI.cpp
between transfer functions and call handling.  Splitting them out
makes both files easier to navigate.

Headers and class layouts unchanged; just a translation-unit move.
