+++ svf/lib/AE/Svfexe/AbstractInterpretation.cpp
@@ -681,71 +681,6 @@ void AbstractInterpretation::handleExtCall(const CallICFGNode *callNode)
     }
 }
 
-/// Check if a function is recursive (part of a call graph SCC)
-bool AbstractInterpretation::isRecursiveFun(const FunObjVar* fun)
-{
-    return preAnalysis->getPointerAnalysis()->isInRecursion(fun);
-}
-
-/// TOP mode for recursive calls: skip the function body entirely and
-/// conservatively set all reachable stores and the return value to TOP.
-void AbstractInterpretation::skipRecursionWithTop(const CallICFGNode *callNode)
-{
-    const RetICFGNode *retNode = callNode->getRetICFGNode();
-
-    // 1. Set return value to TOP
-    if (retNode->getSVFStmts().size() > 0)
-    {
-        if (const RetPE *retPE = SVFUtil::dyn_cast<RetPE>(*retNode->getSVFStmts().begin()))
-        {
-            if (!retPE->getLHSVar()->isPointer() &&
-                    !retPE->getLHSVar()->isConstDataOrAggDataButNotNullPtr())
-                updateAbsValue(retPE->getLHSVar(), IntervalValue::top(), callNode);
-        }
-    }
-
-    // 2. Set all stores in callee's reachable BBs to TOP
-    if (retNode->getOutEdges().size() > 1)
-    {
-        updateAbsState(retNode, getAbsState(callNode));
-        return;
-    }
-    for (const SVFBasicBlock* bb : callNode->getCalledFunction()->getReachableBBs())
-    {
-        for (const ICFGNode* node : bb->getICFGNodeList())
-        {
-            for (const SVFStmt* stmt : node->getSVFStmts())
-            {
-                if (const StoreStmt* store = SVFUtil::dyn_cast<StoreStmt>(stmt))
-                {
-                    const SVFVar* rhsVar = store->getRHSVar();
-                    if (!rhsVar->isPointer() && !rhsVar->isConstDataOrAggDataButNotNullPtr())
-                    {
-                        const AbstractValue& addrs = getAbsValue(store->getLHSVar(), callNode);
-                        if (addrs.isAddr())
-                        {
-                            AbstractState& as = getAbsState(callNode);
-                            for (const auto& addr : addrs.getAddrs())
-                                as.store(addr, IntervalValue::top());
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    // 3. Copy callNode's state to retNode
-    updateAbsState(retNode, getAbsState(callNode));
-}
-
-/// Check if caller and callee are in the same CallGraph SCC (i.e. a recursive callsite)
-bool AbstractInterpretation::isRecursiveCallSite(const CallICFGNode* callNode,
-        const FunObjVar* callee)
-{
-    const FunObjVar* caller = callNode->getCaller();
-    return preAnalysis->getPointerAnalysis()->inSameCallGraphSCC(caller, callee);
-}
-
 /// Get callee function: directly for direct calls, via pointer analysis for indirect calls
 const FunObjVar* AbstractInterpretation::getCallee(const CallICFGNode* callNode)
 {
@@ -772,47 +707,6 @@ const FunObjVar* AbstractInterpretation::getCallee(const CallICFGNode* callNode)
     return SVFUtil::dyn_cast<FunObjVar>(func_var);
 }
 
-/// Skip recursive callsites (within SCC); entry calls from outside SCC are not skipped
-bool AbstractInterpretation::skipRecursiveCall(const CallICFGNode* callNode)
-{
-    const FunObjVar* callee = getCallee(callNode);
-    if (!callee)
-        return false;
-
-    // Non-recursive function: never skip, always inline
-    if (!isRecursiveFun(callee))
-        return false;
-
-    // For recursive functions, skip only recursive callsites (within same SCC).
-    // Entry calls (from outside SCC) are not skipped - they are inlined so that
-    // handleLoopOrRecursion() can analyze the function body.
-    // This applies uniformly to all modes (TOP/WIDEN_ONLY/WIDEN_NARROW).
-    return isRecursiveCallSite(callNode, callee);
-}
-
-/// Check if narrowing should be applied: always for regular loops, mode-dependent for recursion
-bool AbstractInterpretation::shouldApplyNarrowing(const FunObjVar* fun)
-{
-    // Non-recursive functions (regular loops): always apply narrowing
-    if (!isRecursiveFun(fun))
-        return true;
-
-    // Recursive functions: WIDEN_NARROW applies narrowing, WIDEN_ONLY does not
-    // TOP mode exits early in handleLoopOrRecursion, so should not reach here
-    switch (Options::HandleRecur())
-    {
-    case TOP:
-        assert(false && "TOP mode should not reach narrowing phase for recursive functions");
-        return false;
-    case WIDEN_ONLY:
-        return false;  // Skip narrowing for recursive functions
-    case WIDEN_NARROW:
-        return true;   // Apply narrowing for recursive functions
-    default:
-        assert(false && "Unknown recursion handling mode");
-        return false;
-    }
-}
 /// Handle direct or indirect call: get callee(s), process function body, set return state.
 ///
 /// For direct calls, the callee is known statically.
@@ -854,171 +748,8 @@ void AbstractInterpretation::handleFunCall(const CallICFGNode *callNode)
     updateAbsState(retNode, getAbsState(callNode));
 }
 
-/// Handle WTO cycle (loop or recursive function) using widening/narrowing iteration.
-///
-/// Widening is applied at the cycle head to ensure termination of the analysis.
-/// The cycle head's abstract state is iteratively updated until a fixpoint is reached.
-///
-/// == What is being widened ==
-/// The abstract state at the cycle head node, which includes:
-/// - Variable values (intervals) that may change across loop iterations
-/// - For example, a loop counter `i` starting at 0 and incrementing each iteration
-///
-/// == Regular loops (non-recursive functions) ==
-/// All modes (TOP/WIDEN_ONLY/WIDEN_NARROW) behave the same for regular loops:
-/// 1. Widening phase: Iterate until the cycle head state stabilizes
-///    Example: for(i=0; i<100; i++) -> i widens to [0, +inf]
-/// 2. Narrowing phase: Refine the over-approximation from widening
-///    Example: [0, +inf] narrows to [0, 100] using loop condition
-///
-/// == Recursive function cycles ==
-/// Behavior depends on Options::HandleRecur():
-///
-/// - TOP mode:
-///     Does not iterate. Calls skipRecursionWithTop() to set all stores and
-///     return value to TOP immediately. This is the most conservative but fastest.
-///     Example:
-///       int factorial(int n) { return n <= 1 ? 1 : n * factorial(n-1); }
-///       factorial(5) -> returns [-inf, +inf]
-///
-/// - WIDEN_ONLY mode:
-///     Widening phase only, no narrowing for recursive functions.
-///     The recursive function body is analyzed with widening until fixpoint.
-///     Example:
-///       int factorial(int n) { return n <= 1 ? 1 : n * factorial(n-1); }
-///       factorial(5) -> returns [10000, +inf] (widened upper bound)
-///
-/// - WIDEN_NARROW mode:
-///     Both widening and narrowing phases for recursive functions.
-///     After widening reaches fixpoint, narrowing refines the result.
-///     Example:
-///       int factorial(int n) { return n <= 1 ? 1 : n * factorial(n-1); }
-///       factorial(5) -> returns [10000, 10000] (precise after narrowing)
-
-// ---------------------------------------------------------------------------
-// Semi-sparse cycle helpers
-// ---------------------------------------------------------------------------
-//
-// In semi-sparse mode, ValVars live at their def-sites and do not flow
-// through the cycle-head merge. The around-merge widening at cycle_head
-// therefore cannot observe ValVar growth across iterations (e.g. an
-// ArgValVar that a recursive CallPE keeps overwriting at the FunEntry).
-//
-// To fix that, handleLoopOrRecursion runs an extra cross-iter widening on
-// a snapshot that pulls every cycle ValVar into cycle_head's state. The set
-// of ValVar IDs per cycle is precomputed once after AEWTO::initWTO()
-// (initCycleValVars), bottom-up so nested cycles are handled before their
-// enclosing cycle.
-
-// --- Cycle state helpers (dense base) ---
-//
-// The dense default: trace[cycle_head] is the authoritative primary
-// storage, so the snapshot / write-back are trivial.
-// SemiSparseAbstractInterpretation overrides these to additionally pull/scatter
-// cycle ValVars from/to their def-sites.
-
-AbstractState AbstractInterpretation::getFullCycleHeadState(const ICFGCycleWTO* cycle)
-{
-    const ICFGNode* cycle_head = cycle->head()->getICFGNode();
-    AbstractState snap;
-    if (hasAbsState(cycle_head))
-        snap = getAbsState(cycle_head);
-    return snap;
-}
-
-
-bool AbstractInterpretation::widenCycleState(
-    const AbstractState& prev, const AbstractState& cur, const ICFGCycleWTO* cycle)
-{
-    AbstractState prev_copy = prev;
-    AbstractState next = prev_copy.widening(cur);
-    // Always write back (even at fixpoint) so cycle_head's trace holds the
-    // widened state for the upcoming narrowing phase.
-    const ICFGNode* cycle_head = cycle->head()->getICFGNode();
-    abstractTrace[cycle_head] = next;
-    return next == prev;
-}
-
-bool AbstractInterpretation::narrowCycleState(
-    const AbstractState& prev, const AbstractState& cur, const ICFGCycleWTO* cycle)
-{
-    const ICFGNode* cycle_head = cycle->head()->getICFGNode();
-    if (!shouldApplyNarrowing(cycle_head->getFun()))
-        return true;
-    AbstractState prev_copy = prev;
-    AbstractState next = prev_copy.narrowing(cur);
-    if (next == prev)
-        return true;  // fixpoint
-    abstractTrace[cycle_head] = next;
-    return false;
-}
-
-
-void AbstractInterpretation::handleLoopOrRecursion(const ICFGCycleWTO* cycle, const CallICFGNode* caller)
-{
-    const ICFGNode* cycle_head = cycle->head()->getICFGNode();
-
-    // TOP mode for recursive function cycles: set all stores and return value to TOP
-    if (Options::HandleRecur() == TOP && isRecursiveFun(cycle_head->getFun()))
-    {
-        if (caller)
-            skipRecursionWithTop(caller);
-        return;
-    }
-
-    // Iterate until fixpoint with widening/narrowing on the cycle head.
-    bool increasing = true;
-    u32_t widen_delay = Options::WidenDelay();
-    for (u32_t cur_iter = 0;; cur_iter++)
-    {
-        if (cur_iter >= widen_delay)
-        {
-            // getFullCycleHeadState handles dense (returns trace[cycle_head])
-            // and semi-sparse (collects ValVars from def-sites) uniformly.
-            AbstractState prev = getFullCycleHeadState(cycle);
-
-            if (mergeStatesFromPredecessors(cycle_head))
-                handleICFGNode(cycle_head);
-            AbstractState cur = getFullCycleHeadState(cycle);
-
-            if (increasing)
-            {
-                if (widenCycleState(prev, cur, cycle))
-                {
-                    increasing = false;
-                    continue;
-                }
-            }
-            else
-            {
-                if (narrowCycleState(prev, cur, cycle))
-                    break;
-            }
-        }
-        else
-        {
-            // Before widen_delay: process cycle head with gated pattern
-            if (mergeStatesFromPredecessors(cycle_head))
-                handleICFGNode(cycle_head);
-        }
-
-        // Process cycle body components (each with gated merge+handle)
-        for (const ICFGWTOComp* comp : cycle->getWTOComponents())
-        {
-            if (const ICFGSingletonWTO* singleton = SVFUtil::dyn_cast<ICFGSingletonWTO>(comp))
-            {
-                const ICFGNode* node = singleton->getICFGNode();
-                if (mergeStatesFromPredecessors(node))
-                    handleICFGNode(node);
-            }
-            else if (const ICFGCycleWTO* subCycle = SVFUtil::dyn_cast<ICFGCycleWTO>(comp))
-            {
-                if (mergeStatesFromPredecessors(subCycle->head()->getICFGNode()))
-                    handleLoopOrRecursion(subCycle, caller);
-            }
-        }
-    }
-}
+// Loop / recursion handling (handleLoopOrRecursion + cycle helpers +
+// recursion utilities) lives in AELoopRecursion.cpp.
 
 void AbstractInterpretation::handleSVFStatement(const SVFStmt *stmt)
 {
